Push 2- Emilia Emcke
I am struggling a bit with my 5th scene, my egg scene, and my basket scene! I know thats a lot. But I want to be able to make the ball miss the basket. I also want to give the appearance that the egg yolk is popping but I'm having a hard time trying to do that. On the last scene, I want to give the appearance that through the movement of the rectangle (sponge) you are cleaning the screen!